requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
mongoengine==0.20.0
psycopg2==2.8.6
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.2.4
whitenoise==5.2.0
